requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
requests>=2.31.0
google-generativeai>=0.8.3
python-dotenv>=1.0.1
pydantic>=2.6.0
# Updated to allow newer versions compatible with Python 3.13
pillow>=11.0.0